Updating the views, and moved the test.
