+++ templates/QrCodes/view.php
@@ -8,6 +8,7 @@
     <aside class="column">
         <div class="side-nav">
             <h4 class="heading"><?= __('Actions') ?></h4>
+            <?= $this->Html->link(__('Follow'), ['action' => 'forward', $qrCode->qrkey], ['class' => 'side-nav-item']) ?>
             <?= $this->Html->link(__('Edit QR Code'), ['action' => 'edit', $qrCode->id], ['class' => 'side-nav-item']) ?>
             <?= $this->Form->postLink(__('Delete QR Code'), ['action' => 'delete', $qrCode->id], ['confirm' => __('Are you sure you want to delete # {0}?', $qrCode->id), 'class' => 'side-nav-item']) ?>
             <?= $this->Html->link(__('List QR Codes'), ['action' => 'index'], ['class' => 'side-nav-item']) ?>
