+++ tests/TestCase/Controller/QrCodes/ViewTest.php
@@ -54,6 +54,13 @@ class ViewTest extends BaseControllerTest
         $this->get('/qr-codes');
         $this->assertResponseOk();
         $this->helperTestLayoutNormal();
+
+        // test html content.
+        $this->get('/qr-codes');
+        $this->assertResponseOk();
+        $content = (string)$this->_response->getBody();
+        $this->assertSame(1, substr_count($content, '<a href="/qr-codes/view/1">View</a>'));
+        $this->assertSame(1, substr_count($content, '<a href="/qr-codes/forward/sownscribe">Follow</a>'));
     }
 
     /**
@@ -109,6 +116,13 @@ class ViewTest extends BaseControllerTest
         $this->get('/qr-codes/view/1');
         $this->assertResponseOk();
         $this->helperTestLayoutNormal();
+
+        // test html content.
+        $this->get('/qr-codes/view/1');
+        $this->assertResponseOk();
+        $content = (string)$this->_response->getBody();
+        $this->assertSame(1, substr_count($content, '<h3>Sow &amp; Scribe</h3>'));
+        $this->assertSame(1, substr_count($content, '<a href="/qr-codes/forward/sownscribe" class="side-nav-item">Follow</a>'));
     }
 
     /**
